requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.7.2
ccxt==4.3.88